refact: New BaseField class
