+++ src/Form/FieldClass.php
@@ -3,7 +3,7 @@
 namespace Kirby\Form;
 
 use Kirby\Exception\NotFoundException;
-use Kirby\Toolkit\HasI18n;
+use Kirby\Form\Field\BaseField;
 
 /**
  * Abstract field class to be used instead
@@ -18,27 +18,19 @@ use Kirby\Toolkit\HasI18n;
  *
  * @use \Kirby\Cms\HasSiblings<\Kirby\Form\Fields>
  */
-abstract class FieldClass
+abstract class FieldClass extends BaseField
 {
-	use HasI18n;
 	use Mixin\After;
-	use Mixin\Api;
 	use Mixin\Autofocus;
 	use Mixin\Before;
-	use Mixin\DefaultValue;
 	use Mixin\Disabled;
 	use Mixin\Help;
 	use Mixin\Icon;
 	use Mixin\Label;
-	use Mixin\Model;
-	use Mixin\Name;
 	use Mixin\Placeholder;
 	use Mixin\Required;
-	use Mixin\Siblings;
 	use Mixin\Translatable;
 	use Mixin\Validation;
-	use Mixin\Value;
-	use Mixin\When;
 	use Mixin\Width;
 
 	public function __construct(
@@ -57,6 +49,11 @@ abstract class FieldClass
 		array|null $when = null,
 		string|null $width = null
 	) {
+		parent::__construct(
+			name: $name,
+			when: $when
+		);
+
 		$this->setAfter($after);
 		$this->setAutofocus($autofocus);
 		$this->setBefore($before);
@@ -65,11 +62,9 @@ abstract class FieldClass
 		$this->setHelp($help);
 		$this->setIcon($icon);
 		$this->setLabel($label);
-		$this->setName($name);
 		$this->setPlaceholder($placeholder);
 		$this->setRequired($required);
 		$this->setTranslate($translate);
-		$this->setWhen($when);
 		$this->setWidth($width);
 	}
 
@@ -82,60 +77,9 @@ abstract class FieldClass
 		throw new NotFoundException(message: 'Method or option "' . $param . '" does not exist for field type "' . $this->type() . '"');
 	}
 
-	/**
-	 * Returns optional dialog routes for the field
-	 */
-	public function dialogs(): array
+	public function hasValue(): bool
 	{
-		return [];
-	}
-
-	/**
-	 * Returns optional drawer routes for the field
-	 */
-	public function drawers(): array
-	{
-		return [];
-	}
-
-	/**
-	 * Creates a new field instance from a $props array
-	 * @since 6.0.0
-	 */
-	public static function factory(
-		array $props,
-		Fields|null $siblings = null
-	): static {
-		$args = $props;
-
-		unset(
-			$args['model'],
-			$args['type'],
-			$args['value']
-		);
-
-		$field = new static(...$args);
-		$field->setSiblings($siblings);
-
-		if (array_key_exists('model', $props) === true) {
-			$field->setModel($props['model']);
-		}
-
-		if (array_key_exists('value', $props) === true) {
-			$field->fill($props['value']);
-		}
-
-		return $field;
-	}
-
-	public function id(): string
-	{
-		return $this->name();
-	}
-
-	public function isHidden(): bool
-	{
-		return false;
+		return true;
 	}
 
 	/**
@@ -163,47 +107,5 @@ abstract class FieldClass
 			'when'        => $this->when(),
 			'width'       => $this->width(),
 		];
-	}
-
-	/**
-	 * @since 5.2.0
-	 * @todo Move to `Value` mixin once array-based fields are unsupported
-	 */
-	public function reset(): static
-	{
-		$this->value = $this->emptyValue();
-		return $this;
-	}
-
-	/**
-	 * Parses a string template in the given value
-	 */
-	protected function stringTemplate(string|null $string = null): string|null
-	{
-		if ($string !== null) {
-			return $this->model()->toString($string);
-		}
-
-		return null;
-	}
-
-	/**
-	 * Converts the field to a plain array
-	 */
-	public function toArray(): array
-	{
-		$props = $this->props();
-
-		ksort($props);
-
-		return array_filter($props, fn ($item) => $item !== null);
-	}
-
-	/**
-	 * Returns the field type
-	 */
-	public function type(): string
-	{
-		return lcfirst(basename(str_replace(['\\', 'Field'], ['/', ''], static::class)));
 	}
 }
